+++ src/commands/hex-to-text.ts
@@ -1,7 +1,7 @@
 import { window } from 'vscode';
 import { data } from '../data';
 import { insert } from "../utility/editor";
-import { codesToText } from "./code-conversion";
+import { codesToText } from "../utility/code-conversion";
 import { CommandCallback } from './command-callback';
 
 const findEntry = (search: string) => data.find(item => item.codes.length === 1 && item.codes[0] === parseInt(search, 16));