+++ src/commands/identify-characters.ts
@@ -1,7 +1,7 @@
 import { ViewColumn, window, workspace } from 'vscode';
 import { Config } from '../config-interface';
 import { data } from '../data';
-import { codesToHex } from "./code-conversion";
+import { codesToHex } from "../utility/code-conversion";
 import { CommandCallback } from "./command-callback";
 
 export const identifyCharacters: CommandCallback = async (editor) =>